(issue14) Move function_registry to respective app/helpers folder
