+++ app/helpers/agent_registry.py
