+++ tests/app/helpers/function_registry_test.py
@@ -1,13 +1,5 @@
 import pytest
 from app.helpers.function_registry import FunctionRegistry
-import sys
-import os
-
-
-sys.path.insert(
-    0, os.path.abspath(os.path.join(
-        os.path.dirname(__file__), "..", "..", ".."))
-)
 
 
 class TestFunctionRegistry:
